Separate buffer and column implementations
